Refactoring component
